fix: fix some issues
- rename component to GoogleAddressLookup
- fix searchOptions prop
- fix value prop
