+++ src/components/GoogleAddressLookup/__test__/component.spec.js
