+++ src/components/GoogleAddressLookup/icons/locationIcon.js
