+++ src/components/GoogleAddressLookup/icons/locationItemIcon.js
