+++ src/components/GoogleAddressLookup/icons/searchValueIcon.js
