+++ src/components/GoogleAddressLookup/icons/selectedLocationIcon.js
