+++ src/components/GoogleAddressLookup/styles.scss
@@ -38,4 +38,12 @@
             background-color: #f4f6f9;
         }
     }
+
+    .rainbow-google-address-lookup_search-option-label {
+        color: #a4a7b5;
+
+        b {
+            color: #061c3f;
+        }
+    }
 }
